requirements.txt
Pillow
configargparse
python-box
pyyaml
uuid
